--- src/main/java/com/example/FarmMarket/FarmMarketApplication.java
@@ -1,13 +1,0 @@
-package com.example.FarmMarket;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class FarmMarketApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(FarmMarketApplication.class, args);
-	}
-
-}
