--- src/main/java/com/example/FarmMarket/exception/ApplicationException.java
@@ -1,7 +1,0 @@
-package com.example.FarmMarket.exception;
-
-public class ApplicationException extends RuntimeException {
-public ApplicationException (String message) {
-    super(message);
-}
-}
