--- src/main/java/com/example/FarmMarket/exception/ErrorResponse.java
@@ -1,17 +1,0 @@
-package com.example.FarmMarket.exception;
-
-public class ErrorResponse {
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    private String message;
-
-    public ErrorResponse(String message) {
-        this.message = message;
-    }
-}
