--- src/main/java/com/example/FarmMarket/objects/Category.java
@@ -1,50 +1,0 @@
-package com.example.FarmMarket.objects;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-
-public class Category {
-    @Id
-    @GeneratedValue(strategy =
-            GenerationType.IDENTITY)
-    private Integer id;
-    private String categoryName;
-    private String categoryIcon;
-    private String categoryPicture;
-
-    public String getCategoryPicture() {
-        return categoryPicture;
-    }
-
-    public void setCategoryPicture(String categoryPicture) {
-        this.categoryPicture = categoryPicture;
-    }
-
-    public String getCategoryIcon() {
-        return categoryIcon;
-    }
-
-    public void setCategoryIcon(String categoryIcon) {
-        this.categoryIcon = categoryIcon;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getCategoryName() {
-        return categoryName;
-    }
-
-    public void setCategoryName(String categoryName) {
-        this.categoryName = categoryName;
-    }
-}
