--- src/main/java/com/example/FarmMarket/objects/PopUpWindow.java
@@ -1,17 +1,0 @@
-package com.example.FarmMarket.objects;
-
-public class PopUpWindow {
-    private String message;
-
-    public PopUpWindow(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
