--- src/main/java/com/example/FarmMarket/objects/Seller1.java
@@ -1,76 +1,0 @@
-package com.example.FarmMarket.objects;
-
-public class Seller1 {
-    private int id;
-    private String Name;
-    private String Email;
-    private String Username;
-    private String Password;
-    private String PersonalInformation;
-    private String Address;
-    private String Phone;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return Name;
-    }
-
-    public void setName(String name) {
-        Name = name;
-    }
-
-    public String getEmail() {
-        return Email;
-    }
-
-    public void setEmail(String email) {
-        Email = email;
-    }
-
-    public String getUsername() {
-        return Username;
-    }
-
-    public void setUsername(String username) {
-        Username = username;
-    }
-
-    public String getPassword() {
-        return Password;
-    }
-
-    public void setPassword(String password) {
-        Password = password;
-    }
-
-    public String getPersonalInformation() {
-        return PersonalInformation;
-    }
-
-    public void setPersonalInformation(String personalInformation) {
-        PersonalInformation = personalInformation;
-    }
-
-    public String getAddress() {
-        return Address;
-    }
-
-    public void setAddress(String address) {
-        Address = address;
-    }
-
-    public String getPhone() {
-        return Phone;
-    }
-
-    public void setPhone(String phone) {
-        Phone = phone;
-    }
-}
