--- src/main/java/com/example/FarmMarket/repository/CategoryRepository.java
@@ -1,7 +1,0 @@
-package com.example.FarmMarket.repository;
-
-import com.example.FarmMarket.objects.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-}
