--- src/main/java/com/example/FarmMarket/repository/ProductRepository.java
@@ -1,9 +1,0 @@
-package com.example.FarmMarket.repository;
-
-import com.example.FarmMarket.objects.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-}
