--- src/main/java/com/example/FarmMarket/repository/SellerRepository.java
@@ -1,7 +1,0 @@
-package com.example.FarmMarket.repository;
-
-import com.example.FarmMarket.objects.Seller;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SellerRepository extends JpaRepository<Seller,Integer> {
-}
