+++ src/main/java/com/example/farmmarketback/FarmMarketApplication.java
@@ -1,0 +1,13 @@
+package com.example.farmmarketback;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FarmMarketApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(FarmMarketApplication.class, args);
+	}
+
+}
