+++ src/main/java/com/example/farmmarketback/exception/ApplicationException.java
@@ -1,0 +1,7 @@
+package com.example.farmmarketback.exception;
+
+public class ApplicationException extends RuntimeException {
+public ApplicationException (String message) {
+    super(message);
+}
+}
