+++ src/main/java/com/example/farmmarketback/exception/ErrorHandler.java
@@ -1,0 +1,23 @@
+package com.example.farmmarketback.exception;
+
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+@ControllerAdvice
+public class ErrorHandler extends ResponseEntityExceptionHandler {
+
+    @ExceptionHandler (ApplicationException.class)
+    public ResponseEntity<Object> handleApplicationException(ApplicationException e) {
+        e.printStackTrace();
+        return new ResponseEntity<>(new ErrorResponse(e.getMessage()), new HttpHeaders(), HttpStatus.BAD_REQUEST);
+    }
+    @ExceptionHandler (Exception.class)
+    public ResponseEntity<Object> handleError(Exception e) {
+        e.printStackTrace();
+        return new ResponseEntity<>(new ErrorResponse("server error"), new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+}
