+++ src/main/java/com/example/farmmarketback/exception/ErrorResponse.java
@@ -1,0 +1,17 @@
+package com.example.farmmarketback.exception;
+
+public class ErrorResponse {
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    private String message;
+
+    public ErrorResponse(String message) {
+        this.message = message;
+    }
+}
