+++ src/main/java/com/example/farmmarketback/objects/Login.java
@@ -1,0 +1,22 @@
+package com.example.farmmarketback.objects;
+
+public class Login {
+    private String username;
+    private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
