+++ src/main/java/com/example/farmmarketback/objects/PopUpWindow.java
@@ -1,0 +1,17 @@
+package com.example.farmmarketback.objects;
+
+public class PopUpWindow {
+    private String message;
+
+    public PopUpWindow(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
