+++ src/main/java/com/example/farmmarketback/repository/CategoryRepository.java
@@ -1,0 +1,9 @@
+package com.example.farmmarketback.repository;
+
+import com.example.farmmarketback.objects.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+
+
+}
