+++ src/main/java/com/example/farmmarketback/repository/SellerRepository.java
@@ -1,0 +1,7 @@
+package com.example.farmmarketback.repository;
+
+import com.example.farmmarketback.objects.Seller_entity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SellerRepository extends JpaRepository<Seller_entity,Integer> {
+}
