+++ src/main/java/com/example/farmmarketback/security/MyUser.java
@@ -1,0 +1,23 @@
+package com.example.farmmarketback.security;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+
+public class MyUser extends User {
+    private Integer id;
+    public MyUser(Integer id, String username, String password, Collection<? extends GrantedAuthority> authorities) {
+        super(username, password, authorities);
+        this.id = id;
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+}
