+++ src/main/java/com/example/farmmarketback/service/CategoryService.java
@@ -1,0 +1,17 @@
+package com.example.farmmarketback.service;
+import com.example.farmmarketback.objects.Category;
+import com.example.farmmarketback.repository.CategoryRepo;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class CategoryService {
+    @Autowired
+    private CategoryRepo categoryRepo;
+
+    public List<Category> getCategory(){
+        return categoryRepo.getCategory();
+    }
+}
