--- src/test/java/com/example/FarmMarket/FarmMarketApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.FarmMarket;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FarmMarketApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
