+++ src/test/java/com/example/farmmarketback/FarmMarketApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.farmmarketback;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FarmMarketApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
